Simplify command registration
