--- fabric/src/main/java/dev/terminalmc/chatnotify/platform/FabricCommandRegistrar.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2025 TerminalMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.terminalmc.chatnotify.platform;
-
-import dev.terminalmc.chatnotify.command.ChatNotifyCommand;
-import dev.terminalmc.chatnotify.platform.services.ICommandRegistrar;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
-
-public class FabricCommandRegistrar implements ICommandRegistrar {
-    
-    @Override
-    public void registerAll() {
-        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryEvent) -> {
-            ChatNotifyCommand.register(dispatcher);
-        });
-    }
-}
