--- fabric/src/main/resources/META-INF/services/dev.terminalmc.chatnotify.platform.services.ICommandRegistrar
@@ -1,1 +1,0 @@
-dev.terminalmc.chatnotify.platform.FabricCommandRegistrar
